+++ src/main/java/lotto/ui/InputView.java
@@ -1,7 +1,7 @@
 /*
  * UI 입력을 담당하는 클래스
  * */
-package lotto.util;
+package lotto.ui;
 
 import java.util.Arrays;
 import java.util.List;
